Fix pylint
