+++ src/ggrc/builder/json.py
@@ -294,7 +294,7 @@ class UpdateAttrHandler(object):
 
   @classmethod
   def callable_property(cls, obj, json_obj, attr_name, class_attr):
-    """Translate the JSON value for an attr decorated with @colable_property."""
+    """Translate the JSON value for attrs decorated by @colable_property."""
     return json_obj.get(attr_name)()
