+++ src/ggrc/models/hooks/proposal.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2017 Google Inc.
+# Copyright (C) 2018 Google Inc.
 # Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 
 """AccessControlList creation hooks."""
@@ -21,6 +21,7 @@ def is_status_changed_to(required_status, obj):
 
 
 def add_comment_about(proposal, reason, txt):
+  """Create comment about proposal for reason with required text."""
   if not isinstance(proposal.instance, comment.Commentable):
     return
   txt = txt or ""
@@ -39,10 +40,14 @@ def add_comment_about(proposal, reason, txt):
       source=proposal.instance,
       destination=created_comment)
 
+# pylint: disable=unused-argument
+# pylint: disable=too-many-arguments
+
 
 def apply_proposal(
     sender, obj=None, src=None, service=None,
     event=None, initial_state=None):  # noqa
+  """Apply proposal procedure hook."""
   if not is_status_changed_to(obj.STATES.APPLIED, obj):
     return
   obj.applied_by = login.get_current_user()
@@ -57,6 +62,7 @@ def apply_proposal(
 def decline_proposal(
     sender, obj=None, src=None, service=None,
     event=None, initial_state=None):  # noqa
+  """Decline proposal procedure hook."""
   if not is_status_changed_to(obj.STATES.DECLINED, obj):
     return
   obj.declined_by = login.get_current_user()
@@ -67,19 +73,25 @@ def decline_proposal(
 def make_proposal(
     sender, obj=None, src=None, service=None,
     event=None, initial_state=None):  # noqa
+  """Make proposal procedure hook."""
   obj.proposed_by = login.get_current_user()
   add_comment_about(obj, obj.STATES.PROPOSED, obj.agenda)
 
 
 def set_permissions(sender, obj=None, src=None, service=None,
                     event=None, initial_state=None):  # noqa
+  """Set permissions to proposals based on instance ACL model."""
   if sender == all_models.Proposal:
     proposal_model.set_acl_to(obj)
   else:
     proposal_model.set_acl_to_all_proposals_for(obj)
 
+# pylint: enable=unused-argument
+# pylint: enable=too-many-arguments
+
 
 def init_hook():
+  """Init proposal signal handlers."""
   signals.Restful.model_posted.connect(make_proposal,
                                        all_models.Proposal,
                                        weak=False)
