+++ src/ggrc/templates/notifications/proposal_digest.html
@@ -1,5 +1,5 @@
 {#
-Copyright (C) 2017 Google Inc.
+Copyright (C) 2018 Google Inc.
 Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 #}
 <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
